% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dim.wkls_proxy}
\alias{dim.wkls_proxy}
\title{Get dimensions}
\usage{
\method{dim}{wkls_proxy}(x)
}
\arguments{
\item{x}{A wkls_proxy object}
}
\value{
An integer vector of length 2 giving rows and columns,
  or c(0L, 0L) for empty chains
}
\description{
Returns the dimensions (rows and columns) of the resolved data
}
