% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.build_list_query}
\alias{.build_list_query}
\title{Build a query for listing entities with deduplication}
\usage{
.build_list_query(
  country_iso,
  region_iso = NULL,
  subtype_filter,
  distinct_col = "name_primary"
)
}
\description{
Build a query for listing entities with deduplication
}
\keyword{internal}
