% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.is_dependency}
\alias{.is_dependency}
\title{Check if a country code represents a dependency}
\usage{
.is_dependency(country_iso)
}
\description{
Check if a country code represents a dependency
}
\keyword{internal}
