% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.make_geometry_method}
\alias{.make_geometry_method}
\title{Create a geometry method closure}
\usage{
.make_geometry_method(chain, expr_template, ...)
}
\description{
Create a geometry method closure
}
\keyword{internal}
