% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Package load hook
}
\keyword{internal}
