% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is.data.frame.wkls_proxy}
\alias{is.data.frame.wkls_proxy}
\title{Check if object is a wkls_proxy}
\usage{
is.data.frame.wkls_proxy(x)
}
\arguments{
\item{x}{A wkls_proxy object}
}
\value{
A logical value: always returns TRUE for wkls_proxy objects
}
\description{
S3 method that identifies wkls_proxy objects as data frames
}
