% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{names.wkls_proxy}
\alias{names.wkls_proxy}
\title{Get column names}
\usage{
\method{names}{wkls_proxy}(x)
}
\arguments{
\item{x}{A wkls_proxy object}
}
\value{
A character vector of column names, typically:
  "id", "country", "region", "subtype", "name"
}
\description{
Returns the column names of the resolved data
}
