% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ncol.wkls_proxy}
\alias{ncol.wkls_proxy}
\title{Get number of columns}
\usage{
ncol.wkls_proxy(x)
}
\arguments{
\item{x}{A wkls_proxy object}
}
\value{
An integer representing the number of columns, or 0L for empty chains
}
\description{
Returns the number of columns in the resolved data
}
