% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nrow.wkls_proxy}
\alias{nrow.wkls_proxy}
\title{Get number of rows}
\usage{
nrow.wkls_proxy(x)
}
\arguments{
\item{x}{A wkls_proxy object}
}
\value{
An integer representing the number of rows, or 0L for empty chains
}
\description{
Returns the number of rows in the resolved data
}
