% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{overture_version}
\alias{overture_version}
\title{Get the Overture Maps dataset version}
\usage{
overture_version()
}
\value{
Character string with version
}
\description{
Get the Overture Maps dataset version
}
