% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{[.wkls_proxy}
\alias{[.wkls_proxy}
\title{Extract rows/columns like a data frame}
\usage{
\method{[}{wkls_proxy}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{A wkls_proxy object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{drop}{Whether to drop dimensions}
}
\value{
A subset of the resolved data frame
}
\description{
Allows subsetting the resolved data
}
